- Resturcturing program pt 2
