+++ V1 AstroSolver/src/transforms/GroundBasedTransforms/boyde_aux/auxilary_phot_boyde_functions.py
