+++ V1 AstroSolver/src/transforms/TrackRateModeTransforms/RunSCLightCurve.py
